requirements.txt
python-telegram-bot==13.15
scikit-learn
nltk
wikipedia
Pillow
opencv-python
pytesseract
gTTS
SpeechRecognition
requests
numpy
